fix roles and permissions: can compile now!

+++ demo/src/main/java/com/example/identityservice/dto/request/ReviewProductRequest.java
@@ -24,7 +24,7 @@ public class ReviewProductRequest {
     Long userId;
     Long productId;
 
-    Integer rating;
+    Double rating;
     String comment;
     LocalDateTime createdAt;
 }
